+++ pb/sf/substreams/v1/substreams.go
@@ -55,11 +55,6 @@ type ModuleOutputData interface {
 func ValidateRequest(req *Request, isSubRequest bool) error {
 	seenStores := map[string]bool{}
 
-	if req.StartBlockNum < 0 {
-		// TODO(abourget): remove this check once we support StartBlockNum being negative
-		return fmt.Errorf("negative start block %d is not accepted", req.StartBlockNum)
-	}
-
 	if req.Modules == nil {
 		return fmt.Errorf("no modules found in request")
 	}