+++ pb/sf/substreams/v1/substreams_test.go
@@ -14,7 +14,7 @@ func Test_ValidateRequest(t *testing.T) {
 		isSubrequest bool
 		expect       error
 	}{
-		{"negative start block num", TestNewRequest(-1), false, fmt.Errorf("negative start block -1 is not accepted")},
+		{"negative start block num", TestNewRequest(-1, withTestOutputModule("output_mod_1", "map", false)), false, nil},
 		{"no modules found in request", &Request{StartBlockNum: 1}, false, fmt.Errorf("no modules found in request")},
 		{"multiple output modules is not accepted", TestNewRequest(1, withTestOutputModules([][]string{{"output_mod_1", "store"}, {"output_mod_1", "kind"}}, true)), false, fmt.Errorf("multiple output modules is not accepted")},
 		{"store output module is accepted for sub-request", TestNewRequest(1, withTestOutputModule("output_mod_1", "store", false)), true, nil},