+++ docs/release-notes/change-log.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+* Added support for resolving a negative start-block on server (also added to run command)
+
 ## [1.0.1](https://github.com/streamingfast/substreams/releases/tag/v1.0.1)
 
 ### Fixed